+++ src/main/java/com/cognite/client/Relationships.java
@@ -18,7 +18,6 @@ package com.cognite.client;
 
 import com.cognite.client.config.ResourceType;
 import com.cognite.client.config.UpsertMode;
-import com.cognite.client.dto.Event;
 import com.cognite.client.dto.Item;
 import com.cognite.client.dto.Relationship;
 import com.cognite.client.servicesV1.ConnectorServiceV1;
@@ -157,10 +156,7 @@ public abstract class Relationships extends ApiBase {
                 .withIdFunction(this::getRelationshipId);
 
         if (getClient().getClientConfig().getUpsertMode() == UpsertMode.REPLACE) {
-            // Just delete and re-create the relationship since there is no internal id to take into account.
-            return upsertItems.upsertViaDeleteAndCreate(relationships).stream()
-                    .map(this::parseRelationship)
-                    .collect(Collectors.toList());
+            upsertItems = upsertItems.withUpdateMappingFunction(this::toRequestReplaceItem);
         }
 
         return upsertItems.upsertViaCreateAndUpdate(relationships).stream()
@@ -219,6 +215,18 @@ public abstract class Relationships extends ApiBase {
     private Map<String, Object> toRequestUpdateItem(Relationship item) {
         try {
             return RelationshipParser.toRequestUpdateItem(item);
+        } catch (Exception e)  {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /*
+    Wrapping the parser because we need to handle the exception--an ugly workaround since lambdas don't
+    deal very well with exceptions.
+     */
+    private Map<String, Object> toRequestReplaceItem(Relationship item) {
+        try {
+            return RelationshipParser.toRequestReplaceItem(item);
         } catch (Exception e)  {
             throw new RuntimeException(e);
         }