+++ src/main/java/com/cognite/client/util/ParseStruct.java
@@ -3,9 +3,13 @@ package com.cognite.client.util;
 import com.google.common.base.Preconditions;
 import com.google.protobuf.Struct;
 import com.google.protobuf.Value;
+import com.google.protobuf.util.Values;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * This class hosts methods to help parse data from {@code Struct} objects. {@code Struct} is used to represent
@@ -14,20 +18,23 @@ import java.util.List;
 public class ParseStruct {
 
     /**
-     * Parses a given node
-     * @param struct
-     * @param path
-     * @param delimiter
-     * @return
+     * Parses a node in the {@code Struct} into a {@code List<String>}. This is convenient in case the node itself,
+     * or one of its parents, is a list.
+     *
+     * @param struct The Struct to parse.
+     * @param path The path of node to parse, separated by period ("."). Ex: "parent.child.grandChild"
+     * @param delimiter The delimiter to use in the resulting String
+     * @return A delimited string representation of parsed Struct
      */
-    public static String parseDelimitedString(Struct struct, String path, String delimiter) {
+    public static String parseStringDelimited(Struct struct, String path, String delimiter) {
         Preconditions.checkNotNull(struct, "Struct cannot be null");
-        Preconditions.checkArgument(null != path && path.isBlank(),
-                "Node cannot be null or empty");
-        Preconditions.checkArgument(null != delimiter && delimiter.isBlank(),
+        Preconditions.checkArgument(null != path && !path.isBlank(),
+                "Path cannot be null or empty");
+        Preconditions.checkArgument(null != delimiter && !delimiter.isBlank(),
                 "Delimiter cannot be null or empty");
 
-        return "";
+        List<String> pathList = Arrays.asList(path.split("\\."));
+        return parseStringList(struct, pathList).stream().collect(Collectors.joining(delimiter));
     }
 
     /**
@@ -35,109 +42,63 @@ public class ParseStruct {
      * or one of its parents, is a list.
      *
      * @param struct The Struct to parse.
-     * @param path The path of node to parse, separated by period ("."). Ex: "parent.child.grandChild"
+     * @param path The path of node to parse, each item in the list is a path component.
      * @return A list of the data matching the path. If no match, then an empty list is returned.
      */
     public static List<String> parseStringList(Struct struct, List<String> path) {
-        Preconditions.checkNotNull(struct, "Struct cannot be null");
-        Preconditions.checkNotNull(path, "Node cannot be null or empty");
+        Preconditions.checkNotNull(struct, "Struct cannot be null.");
+        Preconditions.checkNotNull(path, "Path cannot be null.");
 
-        return Collections.emptyList();
-    }
-
-    private static List<String> parseStringList(Value value, List<String> path) {
-
-        return Collections.emptyList();
-    }
-
-
-    /*
-
-    fields {
-    key: "to_ActiveSystemStatus"
-    value {
-      struct_value {
-        fields {
-          key: "results"
-          value {
-            list_value {
-              values {
-                struct_value {
-                  fields {
-                    key: "IsUserStatus"
-                    value {
-                      string_value: ""
-                    }
-                  }
-                  fields {
-                    key: "StatusCode"
-                    value {
-                      string_value: "I0072"
-                    }
-                  }
-                  fields {
-                    key: "StatusName"
-                    value {
-                      string_value: "Notification completed"
-                    }
-                  }
-                  fields {
-                    key: "StatusObject"
-                    value {
-                      string_value: "QM000010000010"
-                    }
-                  }
-                  fields {
-                    key: "StatusProfile"
-                    value {
-                      string_value: ""
-                    }
-                  }
-                  fields {
-                    key: "StatusSequenceNumber"
-                    value {
-                      string_value: "00"
-                    }
-                  }
-                  fields {
-                    key: "StatusShortName"
-                    value {
-                      string_value: "NOCO"
-                    }
-                  }
-                  fields {
-                    key: "__metadata"
-                    value {
-                      struct_value {
-                        fields {
-                          key: "id"
-                          value {
-                            string_value: "OTIFICATION_SRV/C_StsObjActiv072\')"
-                          }
-                        }
-                        fields {
-                          key: "type"
-                          value {
-                            string_value: "EAM_OBJPG_MAINTNOTIFICATION_SRV.C_StsObjActiveStatusCodeTextType"
-                          }
-                        }
-                        fields {
-                          key: "uri"
-                          value {
-                            string_value: "https://CodeText(StatusObject=\'QM000010000010\',StatusCode=\'I0072\')"
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
+        List<String> returnList = new ArrayList<>();
+        if (path.size() > 0 && struct.containsFields(path.get(0))) {
+            returnList.addAll(parseStringList(
+                    struct.getFieldsOrDefault(path.get(0), Values.of("")),
+                    path.subList(1, path.size())));
         }
-      }
-    }
-  }
 
+        return returnList;
+    }
+
+    /**
+     * Parses a node in the {@code Value} into a {@code List<String>}. This is convenient in case the node itself,
+     * or one of its parents, is a list.
+     *
+     * @see #parseStringList(Struct, List)
+     *
+     * @param value The Value to parse
+     * @param path The path of node to parse, each item in the list is a path component.
+     * @return A list of the data matching the path. If no match, then an empty list is returned.
      */
+    public static List<String> parseStringList(Value value, List<String> path) {
+        Preconditions.checkNotNull(value, "Value cannot be null.");
+        Preconditions.checkNotNull(path, "Path cannot be null.");
+
+        List<String> returnList = new ArrayList<>();
+        if (path.size() > 0) {
+            // We are not at the child node yet, so we have to unwrap and extract the next level.
+            // This is only possible for a Struct or ValueList.
+            if (value.hasStructValue()) {
+                returnList.addAll(parseStringList(value.getStructValue(), path));
+            } else if (value.hasListValue()) {
+                List<Value> valueList = value.getListValue().getValuesList();
+                for (Value element : valueList) {
+                    returnList.addAll(parseStringList(element, path));
+                }
+            }
+        } else if (path.isEmpty()) {
+            // We are at the child node. Must not be null. Handle list separately.
+            if (value.hasListValue()) {
+                List<Value> valueList = value.getListValue().getValuesList();
+                for (Value element : valueList) {
+                    if (!element.hasNullValue()) {
+                        returnList.add(ParseValue.parseString(element));
+                    }
+                }
+            } else if (!value.hasNullValue()) {
+                returnList.add(ParseValue.parseString(value));
+            }
+        }
+
+        return returnList;
+    }
 }